quotepage updation
added email form and updated footer
